add header and clean ui
